Make script little more user friendly

+++ certificate-generator/gencert.sh
@@ -1,6 +1,44 @@
+#!/bin/bash
 # This file generates the keys and certificates used for testing mosquitto.
 # None of the keys are encrypted, so do not just use this script to generate
 # files for your own use.
+
+OPTS=`getopt -o hc:s: --long server:,help,client: -n 'parse-options' -- "$@"`
+
+if [ $? != 0 ] ; then echo "Failed parsing options." >&2 ; exit 1 ; fi
+
+#echo "$OPTS"
+eval set -- "$OPTS"
+
+HELP=false
+
+while true; do
+  case "$1" in
+    -v | --verbose ) VERBOSE=true; shift ;;
+    -h | --help )    HELP=true; shift ;;
+    -c | --client ) client_arg=$2; shift ; shift ;;
+    -s | --server ) server_arg="$2"; shift; shift ;;
+    -- ) shift; break ;;
+    * ) break ;;
+  esac
+done
+
+help()
+{
+
+echo " ./gencert.sh -c <client hostname/ip> -s <mqttbroker hostname/ip>"
+echo " The certificates are generated in the results folder"
+}
+
+if [ $HELP = true ]; then
+	help
+	exit
+fi
+
+if [ -z "${server_arg// }" ] || [ -z "${client_arg// }" ]; then
+	echo "empty server or client"
+	exit
+fi
 
 rm -rf *.crt *.csr *.key rootCA/ signingCA/ *.pem
 
@@ -11,7 +49,8 @@ for a in root signing; do
 	echo 01 > ${a}CA/serial
 	echo 01 > ${a}CA/crlnumber
 done
-rm -rf certs
+rm -rf results
+mkdir results
 
 BASESUBJ="/C=GB/ST=Derbyshire/L=Derby/O=Mosquitto Project/OU=Testing"
 SBASESUBJ="/C=GB/ST=Nottinghamshire/L=Nottingham/O=Server/OU=Production"
@@ -27,14 +66,15 @@ openssl ca -config openssl.cnf -name CA_root -extensions v3_ca -out test-signing
 
 # Valid server key and certificate.
 openssl genrsa -out server.key 1024
-openssl req -new -key server.key -out server.csr -config openssl.cnf -subj "${SBASESUBJ}/CN=mqttserver/"
+openssl req -new -key server.key -out server.csr -config openssl.cnf -subj "${SBASESUBJ}/CN=${server_arg}/"
 openssl ca -config openssl.cnf -name CA_signing -out server.crt -infiles server.csr
 
 # Valid client key and certificate.
 openssl genrsa -out client.key 1024
-openssl req -new -key client.key -out client.csr -config openssl.cnf -subj "${SBASESUBJ}/CN=mqttclient/"
+openssl req -new -key client.key -out client.csr -config openssl.cnf -subj "${SBASESUBJ}/CN=${client_arg}/"
 openssl ca -config openssl.cnf -name CA_signing -out client.crt -infiles client.csr
 
 cat test-signing-ca.crt test-root-ca.crt > all-ca.crt
 cat client.crt client.key all-ca.crt > client.pem
-c_rehash certs
+
+cp all-ca.crt client.key client.crt server.key server.crt results
